Remove the use case for having a connection-wrapping descriptorForField.
Now deserialization is explicitly aware of connections since they're
their own discreet field types.
